+++ lib/internal/bootstrap/node.js
@@ -140,9 +140,13 @@ function startup() {
   }
 
   if (isMainThread) {
-    mainThreadSetup.setupStdio();
+    const { getStdout, getStdin, getStderr } =
+      NativeModule.require('internal/process/stdio').getMainThreadStdio();
+    setupProcessStdio(getStdout, getStdin, getStderr);
   } else {
-    workerThreadSetup.setupStdio();
+    const { getStdout, getStdin, getStderr } =
+      workerThreadSetup.initializeWorkerStdio();
+    setupProcessStdio(getStdout, getStdin, getStderr);
   }
 
   if (global.__coverage__)
@@ -318,8 +322,14 @@ function startup() {
 function startExecution() {
   // This means we are in a Worker context, and any script execution
   // will be directed by the worker module.
-  if (internalBinding('worker').getEnvMessagePort() !== undefined) {
-    NativeModule.require('internal/worker').setupChild();
+  if (!isMainThread) {
+    const workerThreadSetup = NativeModule.require(
+      'internal/process/worker_thread_only'
+    );
+    // Set up the message port and start listening
+    const { workerFatalExeception } = workerThreadSetup.setup();
+    // Overwrite fatalException
+    process._fatalException = workerFatalExeception;
     return;
   }
 
@@ -509,6 +519,31 @@ function setupProcessObject() {
   const origProcProto = Object.getPrototypeOf(process);
   Object.setPrototypeOf(origProcProto, EventEmitter.prototype);
   EventEmitter.call(process);
+}
+
+function setupProcessStdio(getStdout, getStdin, getStderr) {
+  Object.defineProperty(process, 'stdout', {
+    configurable: true,
+    enumerable: true,
+    get: getStdout
+  });
+
+  Object.defineProperty(process, 'stderr', {
+    configurable: true,
+    enumerable: true,
+    get: getStderr
+  });
+
+  Object.defineProperty(process, 'stdin', {
+    configurable: true,
+    enumerable: true,
+    get: getStdin
+  });
+
+  process.openStdin = function() {
+    process.stdin.resume();
+    return process.stdin;
+  };
 }
 
 function setupGlobalVariables() {
