+++ lib/internal/process/main_thread_only.js
@@ -16,15 +16,6 @@ const {
   validateString
 } = require('internal/validators');
 
-const {
-  setupProcessStdio,
-  getMainThreadStdio
-} = require('internal/process/stdio');
-
-function setupStdio() {
-  setupProcessStdio(getMainThreadStdio());
-}
-
 // The execution of this function itself should not cause any side effects.
 function wrapProcessMethods(binding) {
   function chdir(directory) {
@@ -174,7 +165,6 @@ function setupChildProcessIpcChannel() {
 }
 
 module.exports = {
-  setupStdio,
   wrapProcessMethods,
   setupSignalHandlers,
   setupChildProcessIpcChannel,
