+++ lib/internal/process/stdio.js
@@ -1,6 +1,5 @@
 'use strict';
 
-exports.setupProcessStdio = setupProcessStdio;
 exports.getMainThreadStdio = getMainThreadStdio;
 
 function dummyDestroy(err, cb) { cb(err); }
@@ -131,31 +130,6 @@ function getMainThreadStdio() {
     getStdout,
     getStderr,
     getStdin
-  };
-}
-
-function setupProcessStdio({ getStdout, getStdin, getStderr }) {
-  Object.defineProperty(process, 'stdout', {
-    configurable: true,
-    enumerable: true,
-    get: getStdout
-  });
-
-  Object.defineProperty(process, 'stderr', {
-    configurable: true,
-    enumerable: true,
-    get: getStderr
-  });
-
-  Object.defineProperty(process, 'stdin', {
-    configurable: true,
-    enumerable: true,
-    get: getStdin
-  });
-
-  process.openStdin = function() {
-    process.stdin.resume();
-    return process.stdin;
   };
 }
