+++ lib/worker_threads.js
@@ -2,11 +2,14 @@
 
 const {
   isMainThread,
-  MessagePort,
-  MessageChannel,
   threadId,
   Worker
 } = require('internal/worker');
+
+const {
+  MessagePort,
+  MessageChannel
+} = require('internal/worker/io');
 
 module.exports = {
   isMainThread,
